@@ -1,55 +1,60 @@
 package io.awklabs.sorting.countingsort;
 
-import io.awklabs.sorting.SortingAlgorithmTestBase;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import static org.junit.Assert.assertArrayEquals;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import io.awklabs.sorting.SortingAlgorithmTestBase;
 
 public class CountingSortTest extends SortingAlgorithmTestBase {
 
-    @Override
-    public void before() {
-        super.before();
-        sort = factory.getCountingSort();
-    }
+  @Override
+  @BeforeEach
+  public void before() {
+    super.before();
+    sort = factory.getCountingSort();
+  }
 
-    @Override
-    public void testIntSort() {
-        //Overriding implementation for Counting Sort as counting sort only works for a specified range.
-        //We have a range from 0 - 9
-        int[] array = {1, 4, 1, 2, 7, 5, 2};
-        int[] expected = {1, 1, 2, 2, 4, 5, 7};
-        sort.sort(array);
-        assertArrayEquals(expected, array);
-    }
+  @Test
+  public void testIntSort() {
+    // Overriding implementation for Counting Sort as counting sort only works for a specified
+    // range.
+    // We have a range from 0 - 9
+    int[] array = {1, 4, 1, 2, 7, 5, 2};
+    int[] expected = {1, 1, 2, 2, 4, 5, 7};
+    sort.sort(array);
+    assertArrayEquals(expected, array);
+  }
 
-    @Test(expected = RuntimeException.class)
-    public void testIntSortGreaterThan128() {
-        int[] array = {19, 94, 100, 128, 87, 59, 32};
-        sort.sort(array);
-    }
+  @Test
+  public void testIntSortGreaterThan128() {
+    int[] array = {19, 94, 100, 128, 87, 59, 32};
+    assertThrows(RuntimeException.class, () -> sort.sort(array));
+  }
 
-    @Override
-    @Test(expected = RuntimeException.class)
-    public void testStringSort() {
-        super.testStringSort();
-    }
+  @Override
+  @Test
+  public void testStringSort() {
+    assertThrows(RuntimeException.class, super::testStringSort);
+  }
 
-    @Override
-    @Test(expected = RuntimeException.class)
-    public void testEmptyStringArray() {
-        super.testEmptyStringArray();
-    }
+  @Override
+  @Test
+  public void testEmptyStringArray() {
+    assertThrows(RuntimeException.class, super::testEmptyStringArray);
+  }
 
-    @Override
-    @Test(expected = RuntimeException.class)
-    public void testNullComparator() {
-        super.testNullComparator();
-    }
+  @Override
+  @Test
+  public void testNullComparator() {
+    assertThrows(RuntimeException.class, () -> sort.sort(new String[] {}, null));
+  }
 
-    @Override
-    @Test(expected = RuntimeException.class)
-    public void testNullArray() {
-        super.testNullArray();
-    }
+  @Override
+  @Test
+  public void testNullArray() {
+    assertThrows(RuntimeException.class, () -> sort.sort(null, String::compareTo));
+  }
 }
