@@ -1,47 +1,45 @@
 package io.awklabs.datastructures.tree;
 
-import io.awklabs.datastructures.queue.Queue;
-
 import java.util.List;
+
+import io.awklabs.datastructures.queue.Queue;
 
 public class BinaryTree<T> {
 
-    public Node<T> insertInOrder(Node<T> root, T element) {
-        if (root == null) {
-            root = new Node<>(element);
-            return root;
-        }
-
-        Queue<Node<T>> queue = new Queue<>();
-        queue.enQueue(root);
-
-        while (queue.size() > 0) {
-            Node<T> node = queue.deQueue();
-
-            if (node.left == null) {
-                node.left = new Node<>(element);
-                break;
-            } else
-                queue.enQueue(node.left);
-
-            if (node.right == null) {
-                node.right = new Node<>(element);
-                break;
-            } else
-                queue.enQueue(node.right);
-        }
-
-        return root;
+  public Node<T> insertInOrder(Node<T> root, T element) {
+    if (root == null) {
+      root = new Node<>(element);
+      return root;
     }
 
-    public void inOrder(Node<T> node, List<T> resultArray) {
+    Queue<Node<T>> queue = new Queue<>();
+    queue.enQueue(root);
 
-        if (node == null) {
-            return;
-        }
+    while (queue.size() > 0) {
+      Node<T> node = queue.deQueue();
 
-        inOrder(node.left, resultArray);
-        resultArray.add(node.element);
-        inOrder(node.right, resultArray);
+      if (node.left == null) {
+        node.left = new Node<>(element);
+        break;
+      } else queue.enQueue(node.left);
+
+      if (node.right == null) {
+        node.right = new Node<>(element);
+        break;
+      } else queue.enQueue(node.right);
     }
+
+    return root;
+  }
+
+  public void inOrder(Node<T> node, List<T> resultArray) {
+
+    if (node == null) {
+      return;
+    }
+
+    inOrder(node.left, resultArray);
+    resultArray.add(node.element);
+    inOrder(node.right, resultArray);
+  }
 }
