@@ -1,5 +1,6 @@
 package io.awklabs.datastructures.tree.rbtree;
 
 public enum Colour {
-    RED, BLACK
+  RED,
+  BLACK
 }
