@@ -4,7 +4,7 @@ import java.util.Comparator;
 
 public interface Search<T> {
 
-    int search(int[] array, int key);
+  int search(int[] array, int key);
 
-    int search(T[] array, T key, Comparator<? super T> c);
+  int search(T[] array, T key, Comparator<? super T> c);
 }
