@@ -5,19 +5,17 @@ import io.awklabs.searching.jumpsearch.JumpSearch;
 
 public class SearchAlgorithmFactory {
 
-    private SearchAlgorithmFactory() {
+  private SearchAlgorithmFactory() {}
 
-    }
+  public static SearchAlgorithmFactory getInstance() {
+    return new SearchAlgorithmFactory();
+  }
 
-    public static SearchAlgorithmFactory getInstance() {
-        return new SearchAlgorithmFactory();
-    }
+  public <T> BinarySearch<T> getBinarySearch() {
+    return new BinarySearch<>();
+  }
 
-    public <T> BinarySearch<T> getBinarySearch() {
-        return new BinarySearch<>();
-    }
-
-    public <T> JumpSearch<T> getJumpSearch() {
-        return new JumpSearch<>();
-    }
+  public <T> JumpSearch<T> getJumpSearch() {
+    return new JumpSearch<>();
+  }
 }
