@@ -4,7 +4,7 @@ import java.util.Comparator;
 
 public interface Sort<T> {
 
-    void sort(int[] array);
+  void sort(int[] array);
 
-    void sort(T[] array, Comparator<? super T> c);
+  void sort(T[] array, Comparator<? super T> c);
 }
