@@ -12,47 +12,45 @@ import io.awklabs.sorting.shellsort.ShellSort;
 
 public class SortingAlgorithmFactory {
 
-    private SortingAlgorithmFactory() {
+  private SortingAlgorithmFactory() {}
 
-    }
+  public static SortingAlgorithmFactory getInstance() {
+    return new SortingAlgorithmFactory();
+  }
 
-    public static SortingAlgorithmFactory getInstance() {
-        return new SortingAlgorithmFactory();
-    }
+  public <T> BubbleSort<T> getBubbleSort() {
+    return new BubbleSort<>();
+  }
 
-    public <T> BubbleSort<T> getBubbleSort() {
-        return new BubbleSort<>();
-    }
+  public <T> InsertionSort<T> getInsertionSort() {
+    return new InsertionSort<>();
+  }
 
-    public <T> InsertionSort<T> getInsertionSort() {
-        return new InsertionSort<>();
-    }
+  public <T> SelectionSort<T> getSelectionSort() {
+    return new SelectionSort<>();
+  }
 
-    public <T> SelectionSort<T> getSelectionSort() {
-        return new SelectionSort<>();
-    }
+  public <T> CountingSort<T> getCountingSort() {
+    return new CountingSort<>();
+  }
 
-    public <T> CountingSort<T> getCountingSort() {
-        return new CountingSort<>();
-    }
+  public <T> MergeSort<T> getMergeSort() {
+    return new MergeSort<>();
+  }
 
-    public <T> MergeSort<T> getMergeSort() {
-        return new MergeSort<>();
-    }
+  public <T> QuickSort<T> getQuickSort() {
+    return new QuickSort<>();
+  }
 
-    public <T> QuickSort<T> getQuickSort() {
-        return new QuickSort<>();
-    }
+  public <T> HeapSort<T> getHeapSort() {
+    return new HeapSort<>();
+  }
 
-    public <T> HeapSort<T> getHeapSort() {
-        return new HeapSort<>();
-    }
+  public <T> RadixSort<T> getRadixSort() {
+    return new RadixSort<>();
+  }
 
-    public <T> RadixSort<T> getRadixSort() {
-        return new RadixSort<>();
-    }
-
-    public <T> ShellSort<T> getShellSort() {
-        return new ShellSort<>();
-    }
+  public <T> ShellSort<T> getShellSort() {
+    return new ShellSort<>();
+  }
 }
