@@ -1,13 +1,14 @@
 package io.awklabs.sorting.bubblesort;
 
+import org.junit.jupiter.api.BeforeEach;
+
 import io.awklabs.sorting.SortingAlgorithmTestBase;
-import org.junit.Before;
 
 public class BubbleSortTest extends SortingAlgorithmTestBase {
 
-    @Before
-    public void before() {
-        super.before();
-        sort = factory.getBubbleSort();
-    }
+  @BeforeEach
+  public void before() {
+    super.before();
+    sort = factory.getBubbleSort();
+  }
 }
