@@ -1,12 +1,15 @@
 package io.awklabs.sorting.selectionsort;
 
+import org.junit.jupiter.api.BeforeEach;
+
 import io.awklabs.sorting.SortingAlgorithmTestBase;
 
 public class SelectionSortTest extends SortingAlgorithmTestBase {
 
-    @Override
-    public void before() {
-        super.before();
-        sort = factory.getSelectionSort();
-    }
+  @Override
+  @BeforeEach
+  public void before() {
+    super.before();
+    sort = factory.getSelectionSort();
+  }
 }
