@@ -1,13 +1,15 @@
 package io.awklabs.sorting.shellsort;
 
+import org.junit.jupiter.api.BeforeEach;
+
 import io.awklabs.sorting.SortingAlgorithmTestBase;
 
 public class ShellSortTest extends SortingAlgorithmTestBase {
 
-    @Override
-    public void before() {
-        super.before();
-        sort = factory.getShellSort();
-    }
-
+  @Override
+  @BeforeEach
+  public void before() {
+    super.before();
+    sort = factory.getShellSort();
+  }
 }
